requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
torch>=1.12.0
transformers>=4.20.0
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-ollama>=0.2.0
openai>=1.6.0
pydantic>=2.0.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.9
tqdm>=4.60.0
matplotlib>=3.5.0
seaborn>=0.11.0 
# Local LLM support
accelerate>=0.20.0
bitsandbytes>=0.41.0
safetensors>=0.3.0
tokenizers>=0.13.0
